fix categories dropdown link
