+++ src/components/CategoryDropdownMenu.js
@@ -1,14 +1,11 @@
+/* eslint-disable react/prop-types */
 import React from "react";
 import { FormattedMessage } from "react-intl";
 import { Link } from "react-router-dom";
+import categories from "../locatlData/categoriesData";
 function CategoryDropdownMenu(props) {
   const { catDropdownRef, catDropDown } = props;
-  const categories = [
-    { id: 1, catName: "menClothing" },
-    { id: 2, catName: "womenClothing" },
-    { id: 3, catName: "jewelry" },
-    { id: 4, catName: "electronics" },
-  ];
+
   return (
     <div
       ref={catDropdownRef}
@@ -22,7 +19,11 @@ function CategoryDropdownMenu(props) {
         categories.map((category, index) => (
           <Link
             key={index}
-            to="/productList"
+            to={{
+              pathname: "/productList",
+              search: `${category.catName}`,
+              title: `${category.title}`,
+            }}
             className={
               "text-gray-800 text-sm py-2 px-4 font-normal block w-full whitespace-no-wrap bg-transparent hover:bg-blue-700 hover:text-white"
             }
