+++ src/locatlData/categoriesData.js
@@ -1,0 +1,31 @@
+const categories = [
+  {
+    id: 1,
+    title: "men clothing",
+    catName: "menClothing",
+    catImg:
+      "https://images.unsplash.com/photo-1446214814726-e6074845b4ce?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=376&q=80",
+  },
+  {
+    id: 2,
+    title: "women clothing",
+    catName: "womenClothing",
+    catImg:
+      "https://images.unsplash.com/photo-1571908598047-9b0049221e4e?ixid=MXwxMjA3fDB8MHxzZWFyY2h8NHx8bGFkaWVzJTIwd2VkZGluZyUyMGluZGlhfGVufDB8fDB8&ixlib=rb-1.2.1&auto=format&fit=crop&w=376&q=80",
+  },
+  {
+    id: 3,
+    title: "jewelry",
+    catName: "jewelry",
+    catImg:
+      "https://images.unsplash.com/photo-1515562141207-7a88fb7ce338?ixid=MXwxMjA3fDB8MHxzZWFyY2h8MTN8fGpld2Vscnl8ZW58MHx8MHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=376&q=80",
+  },
+  {
+    id: 4,
+    title: "electronics",
+    catName: "electronics",
+    catImg:
+      "https://images.unsplash.com/photo-1598832517527-7e5b15ee1b2d?ixid=MXwxMjA3fDB8MHxzZWFyY2h8Mjd8fGVsZWN0cm9uaWN8ZW58MHx8MHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=376&q=80",
+  },
+];
+export default categories;
